# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Quiz API",
    "description": "API для системы онлайн-квизов с поддержкой категорий, вопросов, ответов и попыток прохождения",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "660e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "Программирование"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      },
      "Quiz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "770e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "JavaScript Основы"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Тест на знание основ JavaScript"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "example": "660e8400-e29b-41d4-a716-446655440000"
          },
          "timeLimit": {
            "type": "integer",
            "nullable": true,
            "description": "Время в секундах",
            "example": 1800
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "880e8400-e29b-41d4-a716-446655440000"
          },
          "quizId": {
            "type": "string",
            "format": "uuid",
            "example": "770e8400-e29b-41d4-a716-446655440000"
          },
          "text": {
            "type": "string",
            "example": "Что такое замыкание в JavaScript?"
          },
          "order": {
            "type": "integer",
            "example": 1
          },
          "points": {
            "type": "integer",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Answer"
            }
          }
        }
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "990e8400-e29b-41d4-a716-446655440000"
          },
          "questionId": {
            "type": "string",
            "format": "uuid",
            "example": "880e8400-e29b-41d4-a716-446655440000"
          },
          "text": {
            "type": "string",
            "example": "Функция, которая имеет доступ к переменным внешней функции"
          },
          "isCorrect": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      },
      "Attempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "aa0e8400-e29b-41d4-a716-446655440000"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "quizId": {
            "type": "string",
            "format": "uuid",
            "example": "770e8400-e29b-41d4-a716-446655440000"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2024-01-15T11:00:00.000Z"
          },
          "score": {
            "type": "integer",
            "nullable": true,
            "example": 8
          },
          "maxScore": {
            "type": "integer",
            "nullable": true,
            "example": 10
          },
          "status": {
            "type": "string",
            "enum": ["IN_PROGRESS", "COMPLETED", "ABANDONED"],
            "example": "COMPLETED"
          }
        }
      }
    }
  },
  "paths": {
...
    "/answers": {
      "get": {
        "tags": ["Answers"],
        "summary": "Получить варианты ответа для вопроса",
        "parameters": [
          {
            "name": "questionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "880e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Список вариантов ответа",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Answer"
                  }
                },
                "example": [
                  {
                    "id": "990e8400-e29b-41d4-a716-446655440000",
                    "questionId": "880e8400-e29b-41d4-a716-446655440000",
                    "text": "Функция, которая имеет доступ к переменным внешней функции",
                    "isCorrect": true,
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T10:30:00.000Z"
                  },
                  {
                    "id": "990e8400-e29b-41d4-a716-446655440001",
                    "questionId": "880e8400-e29b-41d4-a716-446655440000",
                    "text": "Функция внутри функции",
                    "isCorrect": false,
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T10:30:00.000Z"
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Answers"],
        "summary": "Добавить вариант ответа",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["questionId", "text", "isCorrect"],
                "properties": {
                  "questionId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "880e8400-e29b-41d4-a716-446655440000"
                  },
                  "text": {
                    "type": "string",
                    "example": "Функция высшего порядка"
                  },
                  "isCorrect": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Вариант ответа успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          }
        }
      }
    },
    "/answers/{id}": {
      "get": {
        "tags": ["Answers"],
        "summary": "Получить вариант ответа по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "990e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Вариант ответа найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "404": {
            "description": "Вариант ответа не найден"
          }
        }
      },
      "put": {
        "tags": ["Answers"],
        "summary": "Обновить вариант ответа",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "990e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "questionId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "880e8400-e29b-41d4-a716-446655440000"
                  },
                  "text": {
                    "type": "string",
                    "example": "Функция с лексическим окружением"
                  },
                  "isCorrect": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Вариант ответа успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          },
          "404": {
            "description": "Вариант ответа не найден"
          }
        }
      },
      "delete": {
        "tags": ["Answers"],
        "summary": "Удалить вариант ответа",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "990e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Вариант ответа успешно удален",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "404": {
            "description": "Вариант ответа не найден"
          }
        }
      }
    },
    "/attempts/start": {
      "post": {
        "tags": ["Attempts"],
        "summary": "Начать попытку прохождения квиза",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quizId"],
                "properties": {
                  "quizId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "770e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Попытка успешно начата",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attempt"
                },
                "example": {
                  "id": "aa0e8400-e29b-41d4-a716-446655440000",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "quizId": "770e8400-e29b-41d4-a716-446655440000",
                  "startedAt": "2024-01-15T10:30:00.000Z",
                  "completedAt": null,
                  "score": null,
                  "maxScore": null,
                  "status": "IN_PROGRESS"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          }
        }
      }
    },
    "/attempts/{id}": {
      "get": {
        "tags": ["Attempts"],
        "summary": "Получить состояние попытки",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "aa0e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Состояние попытки",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Attempt"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "quiz": {
                          "$ref": "#/components/schemas/Quiz"
                        },
                        "attemptAnswers": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "questionId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "answerId": {
                                "type": "string",
                                "format": "uuid",
                                "nullable": true
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "id": "aa0e8400-e29b-41d4-a716-446655440000",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "quizId": "770e8400-e29b-41d4-a716-446655440000",
                  "startedAt": "2024-01-15T10:30:00.000Z",
                  "completedAt": null,
                  "score": null,
                  "maxScore": null,
                  "status": "IN_PROGRESS",
                  "quiz": {
                    "id": "770e8400-e29b-41d4-a716-446655440000",
                    "title": "JavaScript Основы",
                    "description": "Тест на знание основ JavaScript",
                    "timeLimit": 1800
                  },
                  "attemptAnswers": []
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Попытка не найдена"
          }
        }
      }
    },
    "/attempts/{id}/answer": {
      "post": {
        "tags": ["Attempts"],
        "summary": "Отправить ответ на вопрос",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "aa0e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["questionId", "answerId"],
                "properties": {
                  "questionId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "880e8400-e29b-41d4-a716-446655440000"
                  },
                  "answerId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "990e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ответ успешно сохранен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "example": "bb0e8400-e29b-41d4-a716-446655440000"
                    },
                    "attemptId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "aa0e8400-e29b-41d4-a716-446655440000"
                    },
                    "questionId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "880e8400-e29b-41d4-a716-446655440000"
                    },
                    "answerId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "990e8400-e29b-41d4-a716-446655440000"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-15T10:35:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Попытка не найдена"
          }
        }
      }
    },
    "/attempts/{id}/complete": {
      "post": {
        "tags": ["Attempts"],
        "summary": "Завершить попытку прохождения квиза",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "aa0e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Попытка успешно завершена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attempt"
                },
                "example": {
                  "id": "aa0e8400-e29b-41d4-a716-446655440000",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "quizId": "770e8400-e29b-41d4-a716-446655440000",
                  "startedAt": "2024-01-15T10:30:00.000Z",
                  "completedAt": "2024-01-15T11:00:00.000Z",
                  "score": 8,
                  "maxScore": 10,
                  "status": "COMPLETED"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Попытка не найдена"
          }
        }
      }
    },
    "/attempts/{id}/result": {
      "get": {
        "tags": ["Attempts"],
        "summary": "Получить результат попытки",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "aa0e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Результат попытки",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Attempt"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "quiz": {
                          "$ref": "#/components/schemas/Quiz"
                        },
                        "attemptAnswers": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "questionId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "answerId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "isCorrect": {
                                "type": "boolean"
                              },
                              "question": {
                                "$ref": "#/components/schemas/Question"
                              },
                              "answer": {
                                "$ref": "#/components/schemas/Answer"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "id": "aa0e8400-e29b-41d4-a716-446655440000",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "quizId": "770e8400-e29b-41d4-a716-446655440000",
                  "startedAt": "2024-01-15T10:30:00.000Z",
                  "completedAt": "2024-01-15T11:00:00.000Z",
                  "score": 8,
                  "maxScore": 10,
                  "status": "COMPLETED",
                  "quiz": {
                    "id": "770e8400-e29b-41d4-a716-446655440000",
                    "title": "JavaScript Основы",
                    "description": "Тест на знание основ JavaScript"
                  },
                  "attemptAnswers": [
                    {
                      "questionId": "880e8400-e29b-41d4-a716-446655440000",
                      "answerId": "990e8400-e29b-41d4-a716-446655440000",
                      "isCorrect": true,
                      "question": {
                        "text": "Что такое замыкание в JavaScript?",
                        "points": 1
                      },
                      "answer": {
                        "text": "Функция, которая имеет доступ к переменным внешней функции"
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Попытка не найдена"
          }
        }
      }
    }
  },
...
}